+++ src/main/java/duke/AddCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.ArrayList;
 
 public class AddCommand extends Command{
