+++ src/main/java/duke/CannotWriteException.java
@@ -1,2 +1,4 @@
+package duke;
+
 public class CannotWriteException extends Exception{
 }
