+++ src/main/java/duke/DukeException.java
@@ -1,2 +1,4 @@
+package duke;
+
 public class DukeException extends Exception{
 }
