+++ src/main/java/duke/Events.java
@@ -1,10 +1,14 @@
+package duke;
+
 public class Events extends Task {
 
     protected String details;
 
+
     public Events(String description, String details) {
         super(description);
         this.details = details;
+
     }
 
     public String getStatus() {
