+++ src/main/java/duke/ExitCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.ArrayList;
 
 public class ExitCommand extends Command{
