+++ src/main/java/duke/FileLoadException.java
@@ -1,2 +1,4 @@
+package duke;
+
 public class FileLoadException extends Exception{
 }
