+++ src/main/java/duke/printcommentException.java
@@ -1,2 +1,4 @@
+package duke;
+
 public class printcommentException extends Exception{
 }
