+++ src/main/java/duke/type.java
@@ -1,3 +1,5 @@
+package duke;
+
 public enum type {
     EVENTS,TODOS,DEADLINES
 }
